+++ daac/s3-replicator.tf
@@ -27,7 +27,7 @@ locals {
 
 module "s3-replicator" {
 
-  source = "https://github.com/nasa/cumulus/releases/download/v8.1.1/terraform-aws-cumulus-s3-replicator.zip"
+  source = "https://github.com/nasa/cumulus/releases/download/v9.2.0/terraform-aws-cumulus-s3-replicator.zip"
 
   prefix               = local.prefix
   vpc_id               = data.aws_vpc.application_vpcs.id